Tests complete (for now) – here’s hoping it’s finally flawless!

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mel-mora <mel-mora@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/11/09 15:24:35 by mel-mora          #+#    #+#             */
-/*   Updated: 2024/11/12 18:19:59 by mel-mora         ###   ########.fr       */
+/*   Updated: 2024/11/13 07:54:28 by mel-mora         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ void	ft_putstr(char *s, int *resu);
 void	ft_putchar(char c, int *resu);
 void	ft_putnbr_dec(int num, int *resu);
 void	ft_putnbr_unsigned(unsigned int num, int *resu);
-void	ft_putnbr_hexa(unsigned long num, int *resu, const char c);
+void	ft_putnbr_hexa(unsigned int num, int *resu, const char c);
 void	ft_tohexa(unsigned long num, int *resu);
 
 #endif
